finalYearProject/requirements.txt
Django
joblib
numpy
scikit-learn
xgboost
django-tailwind
django-compressor
pandas
djangorestframework
shap
jupyter
